Project 2/requirements.txt
Flask==3.1.2
numpy==2.3.3
Pillow==11.3.0
tensorflow==2.20.0
gunicorn
